UEXT interface code separated and made configurable
