+++ src/output_serial.h
@@ -15,13 +15,9 @@
  */
 
 
-#ifndef OUTPUT_H
-#define OUTPUT_H
+#ifndef OUTPUT_SERIAL_H
+#define OUTPUT_SERIAL_H
 
 void output_serial();
-void output_oled();
-
-void init_doglcd();
-void output_doglcd();
 
 #endif
